deprecate: Rename 'bmf.py' to 'fitting.py'
