+++ qha/fitting.py
