+++ qha/multi_configurations/different_phonon_dos.py
@@ -18,7 +18,7 @@ from scipy.constants import Boltzmann
 from scipy.special import logsumexp
 
 import qha.settings
-from qha.bmf import polynomial_least_square_fitting
+from qha.fitting import polynomial_least_square_fitting
 from qha.grid_interpolation import calc_eulerian_strain
 from qha.single_configuration import free_energy
 from qha.type_aliases import Array4D, Scalar, Vector, Matrix
